% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeFeatures.R
\name{ComputeFeatures}
\alias{ComputeFeatures}
\title{Server Module for Feature Computation}
\usage{
ComputeFeatures(input, output, session, network, settings)
}
\arguments{
\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object.}

\item{network}{A reactive object that holds a TSP instance of class "Network".}

\item{settings}{A reactive object that holds a list including several settings.}
}
\description{
This Module provides functionality for Feature Computation
}
