% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeFeaturesUI.R
\name{ComputeFeaturesUI}
\alias{ComputeFeaturesUI}
\title{UI Module for Feature Computation}
\usage{
ComputeFeaturesUI(id)
}
\arguments{
\item{id}{A namespace for the UI.}
}
\value{
A shiny page.
}
\description{
This Module provides a UI object for Feature Computation
}
