% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateNetwork.R
\name{GenerateNetwork}
\alias{GenerateNetwork}
\title{Server Module for Network Generation}
\usage{
GenerateNetwork(input, output, session, parentSession)
}
\arguments{
\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object.}

\item{parentSession}{Shiny session object of parent session.}
}
\value{
A reactive object of class "Network".
}
\description{
This Module provides server logic for Network Generation
}
