% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateNetworkUI.R
\name{GenerateNetworkUI}
\alias{GenerateNetworkUI}
\title{UI Module for Network Generation}
\usage{
GenerateNetworkUI(id)
}
\arguments{
\item{id}{A namespace for the UI.}
}
\value{
A shiny page.
}
\description{
This Module provides a UI object for Network Generation
}
