% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModifyNetwork.R
\name{ModifyNetwork}
\alias{ModifyNetwork}
\title{Server Module for Network Modification}
\usage{
ModifyNetwork(input, output, session, parentSession, network,
  mutations.all, mutations.params)
}
\arguments{
\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object.}

\item{parentSession}{Shiny session object of parent session.}

\item{network}{A reactive object that holds a TSP instance of class "Network".}

\item{mutations.all}{Helper object that holds mutation names and description texts.}

\item{mutations.params}{Helper object that holds mutation names and param values, i.e. presets, lower and upper constraints.}
}
\value{
A list holding reactive objects for the current collection of mutation operators and the modified network.
}
\description{
This Module provides server logic object for Network Modification.
}
