% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModifyNetworkUI.R
\name{ModifyNetworkUI}
\alias{ModifyNetworkUI}
\title{UI Module for Network Modification}
\usage{
ModifyNetworkUI(id, mutations.all, mutations.params)
}
\arguments{
\item{id}{A namespace for the UI.}

\item{mutations.all}{Helper object that holds mutation names and description texts.}

\item{mutations.params}{Helper object that holds mutation names and param values, i.e. presets, lower and upper constraints.}
}
\value{
A shiny page.
}
\description{
This Module provides a UI object for Network Modification
}
