% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mutation.R
\name{Mutation}
\alias{Mutation}
\title{Server Module for Mutations}
\usage{
Mutation(input, output, session, parentSession, mutation, mutation.name,
  mutation.params)
}
\arguments{
\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object.}

\item{parentSession}{Shiny session object of parent session.}

\item{mutation}{Mutation id.}

\item{mutation.name}{Mutation name.}

\item{mutation.params}{Data Frame including params for this mutations, i.e. preset, lower and upper constraints.}
}
\value{
A list containing reactive objects for current mutation values, a logical whether it is active and the probability.
}
\description{
This Module provides server logic for a single Mutation
}
