% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mutations.all}
\alias{mutations.all}
\title{All available mutations in tspgen package}
\format{A data frame with 11 rows and 5 variables.}
\usage{
mutations.all
}
\description{
Contains all avaibale mutation operators and corresponding description texts.
}
\keyword{datasets}
