% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mutations.params}
\alias{mutations.params}
\title{All mutation parameters}
\format{A data frame with 24 rows and 6 variables.}
\usage{
mutations.params
}
\description{
Contains all paramenters for avaibale mutation operators and corresponding values, i.e. preset, lower and upper constraint.
}
\keyword{datasets}
