% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MutationUI.R
\name{MutationUI}
\alias{MutationUI}
\title{UI Module for Mutations}
\usage{
MutationUI(id, mutation, mutation.name, mutation.call, mutation.text,
  mutation.params)
}
\arguments{
\item{id}{A namespace for the UI.}

\item{mutation}{Mutation id.}

\item{mutation.name}{Mutation name.}

\item{mutation.call}{Mutation call.}

\item{mutation.text}{Mutation text.}

\item{mutation.params}{Data Frame including params for this mutations, i.e. preset, lower and upper constraints.}
}
\value{
A shiny container.
}
\description{
This Module provides a UI object for a single Mutation
}
