% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Settings.R
\name{Settings}
\alias{Settings}
\title{Server Module for Settings}
\usage{
Settings(input, output, session)
}
\arguments{
\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object.}
}
\value{
A reactive object holding the current settings.
}
\description{
This Module provides server logic for Settings
}
