% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SettingsUI.R
\name{SettingsUI}
\alias{SettingsUI}
\title{UI Module for Settings}
\usage{
SettingsUI(id)
}
\arguments{
\item{id}{A namespace for the UI.}
}
\value{
A shiny page.
}
\description{
This Module provides a UI object for Application Settings
}
