% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ViewNetwork.R
\name{ViewNetwork}
\alias{ViewNetwork}
\title{Server Module for Network Plotting}
\usage{
ViewNetwork(input, output, session, parentSession, network, settings)
}
\arguments{
\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object.}

\item{parentSession}{Shiny session object of parent session.}

\item{network}{A reactive object that holds a TSP instance of class "Network".}

\item{settings}{A reactive object that holds a list including several settings.}
}
\description{
This Module provides functionality for Network Plotting
}
