% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ViewNetworkUI.R
\name{ViewNetworkUI}
\alias{ViewNetworkUI}
\title{UI Module for Network Plotting}
\usage{
ViewNetworkUI(id)
}
\arguments{
\item{id}{A namespace for the UI.}
}
\value{
A shiny page.
}
\description{
This Module provides a UI object for Network Plotting
}
